+++ src/policy/effect.cpp
@@ -111,16 +111,16 @@ bool IncrementNumericalEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool IncrementNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state);
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double source_eval = m_named_element->get_element()->evaluate(source_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval < target_eval;
 }
 
 bool IncrementNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches);
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval < target_eval;
@@ -149,16 +149,16 @@ bool IncrementOrUnchangedNumericalEffect::are_equal_impl(const BaseEffect& other
 }
 
 bool IncrementOrUnchangedNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state);
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double source_eval = m_named_element->get_element()->evaluate(source_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval <= target_eval;
 }
 
 bool IncrementOrUnchangedNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches);
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval <= target_eval;
@@ -186,16 +186,16 @@ bool DecrementNumericalEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool DecrementNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state);
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double source_eval = m_named_element->get_element()->evaluate(source_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval > target_eval;
 }
 
 bool DecrementNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches);
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval > target_eval;
@@ -223,16 +223,16 @@ bool DecrementOrUnchangedNumericalEffect::are_equal_impl(const BaseEffect& other
 }
 
 bool DecrementOrUnchangedNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state);
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double source_eval = m_named_element->get_element()->evaluate(source_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval >= target_eval;
 }
 
 bool DecrementOrUnchangedNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches);
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval >= target_eval;
@@ -264,8 +264,8 @@ bool UnchangedNumericalEffect::evaluate(const core::State& source_state, const c
 }
 
 bool UnchangedNumericalEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches);
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval == target_eval;
@@ -293,13 +293,13 @@ bool GreaterNumericalEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool GreaterNumericalEffect::evaluate(const core::State&, const core::State& target_state) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (target_eval == INF) return false;
     return target_eval > 0;
 }
 
 bool GreaterNumericalEffect::evaluate(const core::State&, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (target_eval == INF) return false;
     return target_eval > 0;
 }
@@ -326,13 +326,13 @@ bool EqualNumericalEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool EqualNumericalEffect::evaluate(const core::State&, const core::State& target_state) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state);
+    double target_eval = m_named_element->get_element()->evaluate(target_state);
     //if (target_eval == INF) return false;
     return target_eval == 0;
 }
 
 bool EqualNumericalEffect::evaluate(const core::State&, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches);
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches);
     //if (target_eval == INF) return false;
     return target_eval == 0;
 }
@@ -359,16 +359,16 @@ bool IncrementConceptEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool IncrementConceptEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state).size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state).size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state).size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state).size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval < target_eval;
 }
 
 bool IncrementConceptEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval < target_eval;
@@ -396,16 +396,16 @@ bool DecrementConceptEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool DecrementConceptEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state).size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state).size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state).size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state).size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval > target_eval;
 }
 
 bool DecrementConceptEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval > target_eval;
@@ -433,16 +433,16 @@ bool UnchangedConceptEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool UnchangedConceptEffect::evaluate(const core::State& source_state, const core::State& target_state) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state).size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state).size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state).size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state).size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval == target_eval;
 }
 
 bool UnchangedConceptEffect::evaluate(const core::State& source_state, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
+    double source_eval = m_named_element->get_element()->evaluate(source_state, caches)->size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
     //if (source_eval == INF) return false;
     //if (target_eval == INF) return false;
     return source_eval == target_eval;
@@ -470,13 +470,13 @@ bool GreaterConceptEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool GreaterConceptEffect::evaluate(const core::State&, const core::State& target_state) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state).size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state).size();
     //if (target_eval == INF) return false;
     return target_eval > 0;
 }
 
 bool GreaterConceptEffect::evaluate(const core::State&, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
     //if (target_eval == INF) return false;
     return target_eval > 0;
 }
@@ -503,13 +503,13 @@ bool EqualConceptEffect::are_equal_impl(const BaseEffect& other) const {
 }
 
 bool EqualConceptEffect::evaluate(const core::State&, const core::State& target_state) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state).size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state).size();
     //if (target_eval == INF) return false;
     return target_eval == 0;
 }
 
 bool EqualConceptEffect::evaluate(const core::State&, const core::State& target_state, core::DenotationsCaches& caches) const {
-    int target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
+    double target_eval = m_named_element->get_element()->evaluate(target_state, caches)->size();
     //if (target_eval == INF) return false;
     return target_eval == 0;
 }
